fl_client/requirements.txt
# Core cryptographic dependencies
tinyec==0.4.0
ecdsa>=0.18.0

# Machine learning
torch>=2.0.0
numpy>=1.24

# HTTP and environment
requests>=2.31
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0

# Note: web3 dependencies removed to avoid compilation issues
# Install web3 separately if needed: pip install web3 --no-deps
